- Removed Magic Rod's cast bar. - Corrected warning on login sql. - Gravitation no longer causes damage flinch.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@10038 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/login_sql/login.c
@@ -1545,7 +1545,7 @@ int parse_login(int fd)
 					default : error = "Unknown Error."; break;
 					}
 
-					sprintf(tmpsql, "INSERT DELAYED INTO `%s`(`time`,`ip`,`user`,`rcode`,`log`) VALUES (NOW(), '%lu', '%s', '%d','login failed : %s')", loginlog_db, ntohl(ipl), t_uid, result, error);
+					sprintf(tmpsql, "INSERT DELAYED INTO `%s`(`time`,`ip`,`user`,`rcode`,`log`) VALUES (NOW(), '%lu', '%s', '%d','login failed : %s')", loginlog_db, (unsigned int)ntohl(ipl), t_uid, result, error);
 
 					//query
 					if(mysql_query(&mysql_handle, tmpsql)) {
